reorganize code into smaller analyzer units
